+++ src/app/store.ts
@@ -3,7 +3,7 @@ import { currentTodoSlice } from '../features/currentTodo';
 import { todosSlice } from '../features/todos';
 import { filterSlice } from '../features/filter';
 
-const rootReducer = combineSlices( currentTodoSlice, todosSlice, filterSlice );
+const rootReducer = combineSlices(currentTodoSlice, todosSlice, filterSlice);
 
 export const store: EnhancedStore = configureStore({
   reducer: rootReducer,
